requirements.txt
openai-whisper
fastapi
uvicorn
pydantic
python-multipart
argparse
python-dotenv
python-magic
slowapi
librosa